Move `Position` and `Location` to `gradient` module
